# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MEDITRON · СМ-Клиника API
  description: Выявление клинически значимых находок и маршрутизация пациента
  version: 0.0.1
servers:
- url: /
  description: Текущий backend или nginx; /api/sim/* доступен только при ROUTES_SIM_ENABLED=true
...
paths:
...
  /api/routes/{routeId}/notifications:
    get:
      tags:
      - Маршруты
      summary: 'Уведомления маршрута: отправка, прочтение, запись'
      operationId: routeNotifications
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDto'
    post:
      tags:
      - Маршруты
      summary: Ручная отправка по шаблону. Без confirm при сообщении за сутки — 409
        CONFIRM_REQUIRED
      operationId: send
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualNotificationRequest'
        required: true
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NotificationDto'
...
  /api/patients/{patientId}/findings/confirm:
    post:
      tags:
      - Patients
      summary: «Всё ок» — подтвердить пачку находок
      operationId: confirm
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmFindingsRequest'
        required: true
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FindingDto'
...
  /api/routes:
    get:
      tags:
      - Маршруты
      summary: Список маршрутов с фильтрами
      operationId: routes
      parameters:
      - name: stage
        in: query
        required: false
        schema:
          type: string
      - name: specialty
        in: query
        required: false
        schema:
          type: string
      - name: overdue
        in: query
        required: false
        schema:
          type: boolean
      - name: open
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteDto'
  /api/routes/{routeId}:
    get:
      tags:
      - Маршруты
      summary: 'Маршрут подробно: этап, срок, дата посещения или просрочка, находки,
        история этапов'
      operationId: route
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RouteDto'
  /api/route-templates:
    get:
      tags:
      - Маршруты
      summary: Шаблоны маршрутов из словаря (только чтение)
      operationId: routeTemplates
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/patients/{patientId}/routes:
    get:
      tags:
      - Маршруты
      summary: 'Маршруты пациента: сначала открытые'
      operationId: patientRoutes
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteDto'
  /api/patients/{patientId}/notifications:
    get:
      tags:
      - Маршруты
      summary: Все уведомления пациента
      operationId: patientNotifications
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDto'
...
  /api/routes/{routeId}/notification-preview:
    get:
      tags:
      - Маршруты
      summary: Предпросмотр сообщения по шаблону без отправки и сохранения
      operationId: preview
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
      - name: templateCode
        in: query
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NotificationPreviewDto'
  /api/notification-templates:
    get:
      tags:
      - Маршруты
      summary: Шаблоны сообщений; manual=true — доступны для ручной отправки
      operationId: templates
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateDto'
...
  /api/dashboard/patients:
    get:
      tags:
      - Координатор
      summary: 'Переход к пациентам: step — шаг воронки, loss — вид потерь'
      operationId: patients
      parameters:
      - name: step
        in: query
        required: false
        schema:
          type: string
      - name: loss
        in: query
        required: false
        schema:
          type: string
      - name: dateFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: studyType
        in: query
        required: false
        schema:
          type: string
      - name: routeType
        in: query
        required: false
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    TaskDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        role:
          type: string
        patientId:
          type: string
        routeId:
          type: string
        escalationId:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        dueAt:
          type: string
          format: date-time
        overdue:
          type: boolean
        status:
          type: string
        doneBy:
          type: string
        doneAt:
          type: string
          format: date-time
        comment:
          type: string
...
    ManualNotificationRequest:
      type: object
      properties:
        templateCode:
          type: string
          minLength: 1
        confirm:
          type: boolean
        doctor:
          type: string
      required:
      - templateCode
    NotificationDto:
      type: object
      properties:
        id:
          type: string
        routeId:
          type: string
        specialty:
          type: string
        templateCode:
          type: string
        templateTitle:
          type: string
        fullText:
          type: string
        shortText:
          type: string
        links:
          type: object
          additionalProperties:
            type: string
        buttons:
          type: array
          items:
            type: string
        manual:
          type: boolean
        sentBy:
          type: string
        sentAt:
          type: string
          format: date-time
        delivery:
          type: string
        deliveredAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        failedAt:
          type: string
          format: date-time
        bookedAfter:
          type: boolean
        patientId:
          type: string
        patientExternalId:
          type: string
        preferredChannel:
          type: string
...
    EvidenceDto:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        start:
          type: integer
          format: int32
        end:
          type: integer
          format: int32
      required:
      - text
    FindingDto:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        protocolId:
          type: string
        code:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
          - SUGGESTED
          - CONFIRMED
          - REJECTED
          - REMOVED
        source:
          type: string
          enum:
          - ML
          - MANUAL
        evidence:
          $ref: '#/components/schemas/EvidenceDto'
        attributes:
          type: object
          additionalProperties: {}
        confidence:
          type: number
          format: double
        modelVersion:
          type: string
        ruleVersion:
          type: string
        matchedRule:
          type: string
        targetSpecialty:
          type: string
        targetDays:
          type: integer
          format: int32
        level:
          type: string
          enum:
          - EMERGENCY
          - URGENT
          - PLANNED
        urgent:
          type: boolean
        flags:
          type: array
          items:
            type: object
            additionalProperties: {}
        routeId:
          type: string
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        comment:
          type: string
    ConfirmFindingsRequest:
      type: object
      properties:
        findingIds:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
        doctor:
          type: string
      required:
      - findingIds
...
    AppointmentDto:
      type: object
      properties:
        id:
          type: string
        purpose:
          type: string
        dateTime:
          type: string
          format: date-time
        location:
          type: string
        online:
          type: boolean
        doctorName:
          type: string
        status:
          type: string
    FindingRef:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        status:
          type: string
    RouteDto:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        protocolId:
          type: string
        specialty:
          type: string
        routeType:
          type: string
        chainType:
          type: string
        stage:
          type: string
        stageTitle:
          type: string
        pendingVisit:
          type: string
        open:
          type: boolean
        closeReason:
          type: string
        dueAt:
          type: string
          format: date-time
        controlAt:
          type: string
          format: date-time
        visitAt:
          type: string
          format: date-time
        overdueDays:
          type: integer
          format: int32
        slaOverdue:
          type: boolean
        noShowCount:
          type: integer
          format: int32
        tactic:
          type: string
        tacticSubtype:
          type: string
        tacticComment:
          type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/FindingRef'
        stageHistory:
          type: array
          items:
            type: object
            additionalProperties: {}
        openTasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskDto'
        appointment:
          $ref: '#/components/schemas/AppointmentDto'
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        hospitalizationDate:
          type: string
          format: date
        hospitalizationClinic:
          type: string
...
    NotificationPreviewDto:
      type: object
      properties:
        templateCode:
          type: string
        fullText:
          type: string
        shortText:
          type: string
    TemplateDto:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        manual:
          type: boolean
...